[220. Contains Duplicate III][Runtime Error]committed by LeetCode Extension

+++ 220-Contains-Duplicate-III/solution.java
@@ -6,12 +6,12 @@ public class Solution {
         for(int i=0; i<nums.length; i++)
         {
             long reMapNum = (long)nums[i]-Integer.MIN_VALUE;
-            long bucket = reMapNum/(t+1);
+            long bucket = reMapNum/((long)t+1);
             if(bucketMap.containsKey(bucket) || (bucketMap.containsKey(bucket-1) && reMapNum-bucketMap.get(bucket-1)<=t) || (bucketMap.containsKey(bucket+1) && bucketMap.get(bucket-1)-reMapNum<=t))
                 return true;
             if(bucketMap.size()>=k)
             {
-                long lastOne = ((long)nums[i-k]-Integer.MIN_VALUE)/(t+1);
+                long lastOne = ((long)nums[i-k]-Integer.MIN_VALUE)/((long)t+1);
                 bucketMap.remove(lastOne);
             }
             bucketMap.put(bucket, reMapNum);
